Fixed bug #6, explicitly set the QT platform for the shell
https://github.com/flathub/com.valvesoftware.SteamLink/issues/6

When a user with QT_QPA_PLATFORM=wayland in their environment runs this application, it crashes with qt.qpa.plugin: Could not find the Qt platform plugin "wayland" in "". It might be a good idea to set this environment variable to xcb as in https://github.com/flathub/org.keepassxc.KeePassXC/pull/71/files.