Only share X11 display with the app if Wayland is unavailable
This improves the isolation between the app's sandbox and the host
system when running in a Wayland-capable environment, such as
GNOME Shell in its default Wayland mode (tested on Debian 11) or
KDE Plasma Desktop in its non-default Wayland mode (not tested).

The UI shell uses Qt, which was given Wayland support in the previous
commit.

The actual streaming client uses SDL, which defaults to X11 but will
use Wayland if no X11 display is available.

To test:

    Xephyr :2 &
    DISPLAY=:2 flatpak run com.valvesoftware.SteamLink//master

You should see that the Steam Link app (both the shell and the actual
streaming client) continues to run full-screen rather than being confined
to the Xephyr display, indicating that both parts are using Wayland
rather than X11.

Signed-off-by: Simon McVittie <smcv@collabora.com>